moving common api auth fns into a trait & converting clients
