+++ crates/auth_core/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
+async-trait = "0.1.62"
 bytes = "1.2"
 chrono = "0.4"
 dotenv = "0.15"
