+++ crates/google/Cargo.toml
@@ -7,11 +7,12 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
+async-trait = "0.1.62"
+auth_core = { path = "../auth_core" }
 bytes = "1.2"
 chrono = "0.4"
 dotenv = "0.15"
 dotenv_codegen = "0.15"
-auth_core = { path = "../auth_core" }
 log = "0.4"
 oauth2 = "4.2.3"
 reqwest = { version = "0.11", features = ["json"] }
