+++ crates/google/src/lib.rs
@@ -4,6 +4,7 @@ use serde::de::DeserializeOwned;
 use std::str::FromStr;
 
 use anyhow::{anyhow, Result};
+use async_trait::async_trait;
 use oauth2::basic::BasicClient;
 use oauth2::TokenResponse;
 use reqwest::Client;
@@ -12,10 +13,13 @@ use oauth2::basic::BasicTokenResponse;
 use oauth2::reqwest::async_http_client;
 use oauth2::{AuthorizationCode, CsrfToken, PkceCodeChallenge, PkceCodeVerifier, Scope};
 
-pub use libauth::{auth_http_client, oauth_client, AuthorizationRequest, Credentials};
+pub use libauth::{
+    auth_http_client, oauth_client, ApiClient, AuthorizationRequest, Credentials, OnRefreshFn,
+};
+
 pub mod types;
 use types::{
-    AuthScope, CalendarEvent, CalendarListResponse, File, FileType, Files, GoogUser,
+    CalendarEvent, CalendarListResponse, File, FileType, Files, GoogUser,
     ListCalendarEventsResponse,
 };
 
@@ -28,14 +32,70 @@ const AUTH_URL: &str = "https://accounts.google.com/o/oauth2/v2/auth";
 const TOKEN_URL: &str = "https://www.googleapis.com/oauth2/v3/token";
 const REVOKE_URL: &str = "https://oauth2.googleapis.com/revoke";
 
-pub type OnRefreshFn = Box<dyn FnMut(&Credentials) + Send + Sync + 'static>;
-
 pub struct GoogClient {
     endpoint: String,
     http: Client,
     pub oauth: BasicClient,
     pub credentials: Credentials,
     pub on_refresh: OnRefreshFn,
+}
+
+#[async_trait]
+impl ApiClient for GoogClient {
+    fn id(&self) -> String {
+        self.endpoint.clone()
+    }
+
+    fn set_credentials(&mut self, credentials: &Credentials) -> Result<()> {
+        self.credentials = credentials.clone();
+        self.http = auth_http_client(credentials.access_token.secret())?;
+        Ok(())
+    }
+
+    fn set_on_refresh(&mut self, callback: impl FnMut(&Credentials) + Send + Sync + 'static) {
+        self.on_refresh = Box::new(callback);
+    }
+
+    fn authorize(&self, scopes: &[String]) -> AuthorizationRequest {
+        // Google supports Proof Key for Code Exchange (PKCE - https://oauth.net/2/pkce/).
+        // Create a PKCE code verifier and SHA-256 encode it as a code challenge.
+        let (pkce_code_challenge, pkce_code_verifier) = PkceCodeChallenge::new_random_sha256();
+
+        let scopes = scopes
+            .iter()
+            .map(|s| Scope::new(s.to_string()))
+            .collect::<Vec<Scope>>();
+
+        // Generate the authorization URL to which we'll redirect the user.
+        let (authorize_url, csrf_state) = self
+            .oauth
+            .authorize_url(CsrfToken::new_random)
+            .add_scopes(scopes)
+            .set_pkce_challenge(pkce_code_challenge.clone())
+            .url();
+
+        AuthorizationRequest {
+            url: authorize_url,
+            csrf_token: csrf_state,
+            pkce_challenge: pkce_code_challenge,
+            pkce_verifier: pkce_code_verifier.secret().to_string(),
+        }
+    }
+
+    async fn token_exchange(&self, code: &str, pkce_verifier: &str) -> Result<BasicTokenResponse> {
+        let code = AuthorizationCode::new(code.to_owned());
+
+        match self
+            .oauth
+            .exchange_code(code)
+            .set_pkce_verifier(PkceCodeVerifier::new(pkce_verifier.to_owned()))
+            .request_async(async_http_client)
+            .await
+        {
+            Ok(val) => Ok(val),
+            Err(err) => Err(anyhow!(err.to_string())),
+        }
+    }
 }
 
 impl GoogClient {
@@ -67,10 +127,6 @@ impl GoogClient {
             credentials: creds,
             on_refresh: Box::new(|_| {}),
         })
-    }
-
-    pub fn set_on_refresh(&mut self, callback: impl FnMut(&Credentials) + Send + Sync + 'static) {
-        self.on_refresh = Box::new(callback);
     }
 
     pub async fn call(
@@ -121,12 +177,6 @@ impl GoogClient {
             (self.on_refresh)(&self.credentials);
         }
 
-        Ok(())
-    }
-
-    pub fn set_credentials(&mut self, credentials: &Credentials) -> Result<()> {
-        self.credentials = credentials.clone();
-        self.http = auth_http_client(credentials.access_token.secret())?;
         Ok(())
     }
 
@@ -269,50 +319,5 @@ impl GoogClient {
     pub async fn get_user(&mut self) -> Result<GoogUser> {
         let endpoint = "https://www.googleapis.com/oauth2/v3/userinfo";
         self.call_json(endpoint, &Vec::new()).await
-    }
-
-    pub fn authorize(&self, scopes: &[AuthScope]) -> AuthorizationRequest {
-        // Google supports Proof Key for Code Exchange (PKCE - https://oauth.net/2/pkce/).
-        // Create a PKCE code verifier and SHA-256 encode it as a code challenge.
-        let (pkce_code_challenge, pkce_code_verifier) = PkceCodeChallenge::new_random_sha256();
-
-        let scopes = scopes
-            .iter()
-            .map(|s| Scope::new(s.as_ref().to_string()))
-            .collect::<Vec<Scope>>();
-
-        // Generate the authorization URL to which we'll redirect the user.
-        let (authorize_url, csrf_state) = self
-            .oauth
-            .authorize_url(CsrfToken::new_random)
-            .add_scopes(scopes)
-            .set_pkce_challenge(pkce_code_challenge.clone())
-            .url();
-
-        AuthorizationRequest {
-            url: authorize_url,
-            csrf_token: csrf_state,
-            pkce_challenge: pkce_code_challenge,
-            pkce_verifier: pkce_code_verifier.secret().to_string(),
-        }
-    }
-
-    pub async fn token_exchange(
-        &self,
-        code: &str,
-        pkce_verifier: &str,
-    ) -> Result<BasicTokenResponse> {
-        let code = AuthorizationCode::new(code.to_owned());
-
-        match self
-            .oauth
-            .exchange_code(code)
-            .set_pkce_verifier(PkceCodeVerifier::new(pkce_verifier.to_owned()))
-            .request_async(async_http_client)
-            .await
-        {
-            Ok(val) => Ok(val),
-            Err(err) => Err(anyhow!(err.to_string())),
-        }
     }
 }
